backend/requirements.txt
fastapi==0.95.1
uvicorn==0.15.0
sqlalchemy==1.4.23
pydantic==2.7.0
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.5
psycopg2-binary==2.9.1
alembic==1.7.1